+++ src/main/java/edu/online/messenger/service/UserService.java
@@ -1,7 +1,5 @@
 package edu.online.messenger.service;
 
-import org.springframework.stereotype.Service;
-
 public interface UserService {
 
     boolean existsById(Long id);