+++ src/main/java/edu/online/messenger/controller/UserController.java
@@ -18,4 +18,9 @@ public class UserController {
     public boolean existsById(@PathVariable Long userId) {
         return userService.existsById(userId);
     }
+
+    @GetMapping("/existence/login/{login}")
+    public boolean existsByLogin(@PathVariable String login) {
+        return userService.existsByLogin(login);
+    }
 }