+++ src/main/java/edu/online/messenger/service/UserService.java
@@ -1,9 +1,8 @@
 package edu.online.messenger.service;
 
-import org.springframework.stereotype.Service;
-
 public interface UserService {
 
     boolean existsById(Long id);
+    boolean existsByLogin(String login);
 }