+++ src/main/java/edu/online/messenger/service/impl/UserServiceImpl.java
@@ -2,8 +2,8 @@ package edu.online.messenger.service.impl;
 
 import edu.online.messenger.repository.UserRepository;
 import edu.online.messenger.service.UserService;
-import org.springframework.stereotype.Service;
 import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Service;
 
 @Service
 @AllArgsConstructor
@@ -14,5 +14,10 @@ public class UserServiceImpl implements UserService {
     @Override
     public boolean existsById(Long id) {
         return repository.existsById(id);
+    }
+
+    @Override
+    public boolean existsByLogin(String login) {
+        return repository.existsByLogin(login);
     }
 }