+++ CHANGELOG.md
@@ -43,7 +43,7 @@
 Добавлена модель User
 
 ### version 4.0
-Разработаны UserController, UserService, UserRepository. Создан следующий эндпойнт:
+Разработаны UserController, UserService, UserRepository. Создан следующий эндпоинт:
 ## UserController 
 ### Get
 #### Проверка наличия пользователя по id
@@ -53,11 +53,10 @@ Bad request 400, если в качестве user_id переданы не ва
 
 
 ### version 5.0
-Разработаны UserController, UserService, UserRepository. Создан следующий эндпоинт:
+Создан следующий эндпоинт:
 ## UserController
 ### Get
 #### Проверка наличия пользователя по логину
 #### Запрос:http://localhost:9087/api/users/existence/login/{login}
 #### Ответ: boolean (true/false)
 Bad request 400, если в качестве user_id переданы не валидные данные
-В случае, если пользователь не существует, никаких исключений не выбрасывается